Добавляет модульность.
